Add a new RecognizerIntent based on VoiceImeView +
refactor the existing RecognizerIntent to share code with the new one.
