+++ app/build.gradle
@@ -24,8 +24,8 @@ android {
         applicationId 'ee.ioc.phon.android.speak'
         minSdkVersion 14
         targetSdkVersion 22
-        versionCode 1206
-        versionName '1.2.06'
+        versionCode 1301
+        versionName '1.3.01'
     }
